@@ -1,19 +1,19 @@
 import React, { useState } from 'react';
-import './login.css';
+import styles from './Login.module.scss';
 
 import { connect } from 'react-redux';
 
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 
-import { signInWithGoogle, signInWithPassword, createUserProfileDocument } from '../../firebase/firebase.utils';
-import { setUser } from '../../redux/user/userActions';
+import { createUserProfileDocument, signInWithGoogle, signInWithPassword } from '../../firebase/firebase.utils';
 import { clearStateAction } from '../../redux/list/listActions';
+import { setUser } from '../../redux/user/userActions';
 
-import Form from 'react-bootstrap/Form';
 import Button from 'react-bootstrap/Button';
+import Form from 'react-bootstrap/Form';
 
 const Login = ({ dispatch }) => {
-  const history = useHistory();
+  const navigate = useNavigate();
   const [validated, setValidated] = useState(false);
   const [loginEmail, setLoginEmail] = useState('');
   const [loginPass, setLoginPass] = useState('');
@@ -40,7 +40,7 @@ const Login = ({ dispatch }) => {
   const logInWithGoogle = async () => {
     const userAuth = await signInWithGoogle();
     if (userAuth) {
-      history.push('/home');
+      navigate('/home');
     }
   };
 
@@ -58,7 +58,7 @@ const Login = ({ dispatch }) => {
           );
         });
 
-        history.push('/home');
+        navigate('/home');
       } else {
         alert('You need to verify your email. Please check your Inbox and follow the link.');
       }
@@ -67,29 +67,36 @@ const Login = ({ dispatch }) => {
 
   return (
     <div className="container">
-      <div className="row login-row">
-        <div className="col-sm-6 login-col">
-          <Form className="login-form justify-content-center" noValidate validated={validated} onSubmit={handleSubmit}>
-            <Form.Group controlId="formBasicEmail">
-              <Form.Label>Email address</Form.Label>
-              <Form.Control type="email" placeholder="Enter email" required onChange={onEmailChange} />
+      <div className={`row ${styles.login_row}`}>
+        <div className={`col-sm-6 ${styles.login_col}`}>
+          <Form
+            className={`${styles.login_form} justify-content-center`}
+            noValidate
+            validated={validated}
+            onSubmit={handleSubmit}
+          >
+            <Form.Group controlId="formBasicEmail" className='pb-2'>
+              <Form.Label className={styles.label}>Email address</Form.Label>
+              <Form.Control className={styles.form_control_custom} type="email" placeholder="Enter email" required onChange={onEmailChange} />
               <Form.Control.Feedback type="invalid">Please type an email address</Form.Control.Feedback>
               <Form.Control.Feedback type="valid">Looks Good</Form.Control.Feedback>
             </Form.Group>
 
-            <Form.Group controlId="formBasicPassword">
-              <Form.Label>Password</Form.Label>
-              <Form.Control type="password" placeholder="Password" required onChange={onPassChange} />
+            <Form.Group controlId="formBasicPassword" className='pb-2'>
+              <Form.Label className={styles.label}>Password</Form.Label>
+              <Form.Control className={styles.form_control_custom} type="password" placeholder="Password" required onChange={onPassChange} />
               <Form.Control.Feedback type="invalid">Please type your password</Form.Control.Feedback>
               <Form.Control.Feedback type="valid">Looks Good</Form.Control.Feedback>
             </Form.Group>
-            <Button onClick={logInWithPassword} variant="primary" type="submit">
+            <Button className={`mt-3 ${styles.btn_primary_custom}`} onClick={logInWithPassword} variant="primary" type="submit" size="sm">
               Log In
             </Button>
           </Form>
-          <button onClick={logInWithGoogle} className="google_signIn">
-            Sing in with Google
-          </button>
+          <Button variant="light" onClick={logInWithGoogle}>
+            <span className="text-danger">Sign-In{" "}</span>
+            <span className="text-primary">with{" "}</span>
+            <span className="text-success">Google</span>
+          </Button>
         </div>
       </div>
     </div>
