@@ -2,7 +2,7 @@ import React from 'react';
 import './LoginPage.css';
 
 import Layout from '../../components/Layout/layout';
-import Login from '../../components/Login/login';
+import Login from '../../components/Login/Login';
 
 const LoginPage = () => {
   return (
