@@ -2,7 +2,7 @@ import React from 'react';
 import './RegisterPage.css';
 
 import Layout from '../../components/Layout/layout';
-import Register from '../../components/Register/register';
+import Register from '../../components/Register/Register';
 
 const RegisterPage = () => {
   return (
